Strip `.sh` from WindGate commands.
