+++ windgate-project/cli/.gitignore
