+++ windgate-project/cli/src/main/assembly/dist.xml
