+++ windgate-project/cli/src/main/resources/com/asakusafw/windgate/cli/log.properties
