+++ windgate-project/cli/src/test/resources/com/asakusafw/windgate/cli/loadProperties.properties
